Refactor the Report class into Response and Request ones.
